pyproject.toml
[build-system]
build-backend = "poetry.masonry.api"
requires = ["poetry>=0.12"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "classes"
version = "0.1.0"
[tool.poetry.dependencies]
dnspython = "^1.16"
pymongo = "^3.10"
python = "^3.7"
[tool.poetry.dev-dependencies]
